test the build tests are not yet working
