+++ circle.yml
@@ -5,12 +5,15 @@ machine:
 dependencies:
   override:
     - docker info
-    - docker-compose -f docker-compose-build.yml up
+    - docker build --build-arg BuildNumber=local -t httpservice-build .
+    - docker create --name httpservice-build httpservice-build
+    - docker cp httpservice-build:/package ./package
+    - docker rm httpservice-build
 
 test:
   override:
     - mkdir -p $CIRCLE_TEST_REPORTS/junit
-    - sudo mv testresults/out.xml $CIRCLE_TEST_REPORTS/junit
+    # - sudo mv testresults/out.xml $CIRCLE_TEST_REPORTS/junit
     
 deployment:
   hub:
