--- docker-compose-build.yml
@@ -1,8 +1,0 @@
-app:
-  image: microsoft/dotnet:1.1.0-sdk-projectjson
-  volumes:
-    - .:/sln
-  working_dir: /sln
-  environment: 
-    - "BuildNumber=${CIRCLE_BUILD_NUM}"
-  entrypoint: ["sh", "./build-container.sh"]
