--- src/HttpService/project.json
@@ -1,13 +1,0 @@
-{
-  "version": "1.1.1",
-  "dependencies": {
-    "Microsoft.AspNetCore.Authentication": "1.1.0",
-    "Microsoft.AspNetCore.Http.Abstractions": "1.1.0",
-    "NETStandard.Library": "1.6.1"
-  },
-  "frameworks": {
-    "netstandard1.6": {
-      "imports": "dnxcore50"
-    }
-  }
-}
