--- test/HttpService.Test/project.json
@@ -1,32 +1,0 @@
-{
-  "version": "1.0.0-*",
-  "buildOptions": {
-    "debugType": "portable"
-  },
-  "dependencies": {
-    "System.Runtime.Serialization.Primitives": "4.3.0",
-    "xunit": "2.2.0-beta2-build3300",
-    "dotnet-test-xunit": "2.2.0-preview2-build1029",
-    "HttpService": {
-      "target": "project"
-    },
-    "Moq": "4.6.38-alpha",
-    "FluentAssertions": "4.9.1",
-    "Microsoft.AspNetCore.TestHost": "1.1.0"
-  },
-  "testRunner": "xunit",
-  "frameworks": {
-    "netcoreapp1.1": {
-      "dependencies": {
-        "Microsoft.NETCore.App": {
-          "type": "platform",
-          "version": "1.1.0"
-        }
-      },
-      "imports": [
-        "dotnet5.4",
-        "portable-net451+win8"
-      ]
-    }
-  }
-}
